requirements.txt
streamlit
Pyrebase4
firebase-admin